requirements.txt
boto3
dbt-core
dbt-redshift
apache-airflow-providers-docker
python-dotenv